+++ src/components/common/PopOverSort/StylePopOverSort.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 export const Container = styled.div`
   width: 79px;
+  margin-top: 4px;
   padding: 4px 0px;
   border-radius: 8px;
   border: 1px solid var(--gray30);
